service/requirements.txt
# Dependências do Streamlit App
streamlit>=1.28.0
tensorflow>=2.15.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
Pillow>=10.0.0
requests>=2.31.0
